Teste2projeto/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
bcrypt==4.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
dataclasses-json==0.6.7
dnspython==2.7.0
exceptiongroup==1.3.0
frozenlist==1.7.0
gensim==4.3.3
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
imbalanced-learn==0.13.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-community==0.2.5
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langsmith==0.1.147
marshmallow==3.26.1
multidict==6.6.3
mypy_extensions==1.1.0
numpy==1.26.4
ollama==0.2.1
orjson==3.11.1
packaging==24.2
pandas==2.3.1
paramiko==3.5.1
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.11
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sklearn-compat==0.1.3
smart_open==7.3.0.post1
sniffio==1.3.1
SQLAlchemy==2.0.42
sshtunnel==0.4.0
tenacity==8.5.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wrapt==1.17.2
yarl==1.20.1
